feed_monitor/requirements.txt
py-IoticAgent==0.5.0
Jinja2==2.9.5
humanize==0.5.1
